Додано частину ДЗ hw3          Задание 2

+++ pavel/src/main/hw/hw2/TaskThree.java
@@ -14,7 +14,7 @@ public class TaskThree {
         int b = 4;
         int c = 3;
         int e = 2;
-        double d = a + b/c/e;
+        double d = (float)a + ((float)b/c/e);
         System.out.println(d);
     }
 }
